+++ CHANGELOG.md
@@ -87,6 +87,7 @@
   into crate `cosmwasm-storage`.
 - Rename `Result` to `StdResult` to differentiate between the auto-`use`d
   `core::result::Result`. Fix error argument to `Error`.
+- Rename `Error` to `StdError`.
 
 **cosmwasm-vm**